+++ CHANGELOG.md
@@ -8,12 +8,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- The next version number should be the version already set in pubspec.yaml -->
 ### Fixed
 - Android: Fix various compilation warnings [#634](https://github.com/rokwire/app-flutter-plugin/issues/634).
+- Fixed text color of ImageUtils.mapMarkerImage, removed "Group" from API name [#5344](https://github.com/rokwire/illinois-app/issues/5344).
 ### Added
 - Added ListUtils.last() helper [#5344](https://github.com/rokwire/illinois-app/issues/5344).
 - Added ListUtils.stripNull() helper [#5344](https://github.com/rokwire/illinois-app/issues/5344).
 - Add file attachments to conversation messages [#639](https://github.com/rokwire/app-flutter-plugin/issues/639).
 ### Changed
-- ImageUtils.mapGroupMarkerImage extended to handle pin markers [#5344](https://github.com/rokwire/illinois-app/issues/5344).
+- ImageUtils.mapGroupMarkerImage extended to handle pin and explore markers [#5344](https://github.com/rokwire/illinois-app/issues/5344).
 
 ## [1.11.2] - 2025-09-18
 ### Changed